(#2)feat/delete board

+++ src/main/java/com/techeer/port/voilio/domain/board/repository/BoardRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardRepository extends JpaRepository<Board, Integer> {}
+public interface BoardRepository extends JpaRepository<Board, Long> {}
